Jd udup 1.0

See merge request universe/udup!15

+++ internal/models/order.go
@@ -23,7 +23,7 @@ type Order struct {
 
 	SkuId string
 
-	TrafficAgainstLimits uint64
+	TrafficAgainstLimits int
 
 	TotalTransferredBytes uint64
